move more common functions into utils
